#243 TECHNICAL CHANGE reorganise settings and migration code into dedicated directories
