+++ src/main/settings/__tests__/settings-json.test.ts
