+++ src/main/settings/__tests__/theme-settings.test.ts
