+++ src/main/settings/settings-json.ts
